+++ sat/cli/bootsys/mgmt_shutdown_ansible.py
@@ -1,0 +1,48 @@
+"""
+Run an Ansible playbook to shutdown platform services on the management cluster.
+
+(C) Copyright 2020 Hewlett Packard Enterprise Development LP.
+
+Permission is hereby granted, free of charge, to any person obtaining a
+copy of this software and associated documentation files (the "Software"),
+to deal in the Software without restriction, including without limitation
+the rights to use, copy, modify, merge, publish, distribute, sublicense,
+and/or sell copies of the Software, and to permit persons to whom the
+Software is furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included
+in all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.  IN NO EVENT SHALL
+THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR
+OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+OTHER DEALINGS IN THE SOFTWARE.
+"""
+import logging
+import shlex
+import subprocess
+
+LOGGER = logging.getLogger(__name__)
+
+PLAYBOOK = '/opt/cray/crayctl/ansible_framework/main/platform-shutdown.yml'
+
+
+def do_shutdown_playbook():
+    cmd = 'ansible-playbook --skip-tags prompt {}'.format(PLAYBOOK)
+    LOGGER.debug('Invoking Ansible: %s', cmd)
+
+    try:
+        proc = subprocess.run(shlex.split(cmd), stdout=subprocess.PIPE,
+                              stderr=subprocess.PIPE, encoding='utf-8')
+    except OSError as err:
+        LOGGER.error("Failed to invoke '%s': %s", cmd, err)
+        raise SystemExit(1)
+
+    if proc.returncode:
+        LOGGER.error("Command '%s' failed. stderr: %s", cmd, proc.stderr)
+        raise SystemExit(1)
+
+    LOGGER.info('Ansible playbook complete.')